Verify the resource metadata
